+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.3.0
+* [Idea card](https://github.com/ZorianaPro/hackathon2020/issues/7)
+####To DO:
+* Add team members component
+* Add functions to open overlays
+* Take all info from the state
+
 ## 1.2.0
 * [Main Section](https://github.com/ZorianaPro/hackathon2020/issues/4)
 * Add button component